backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.3.7
langchain-groq==0.2.4
langchain-core==0.3.19
langgraph==0.2.48
openai==1.54.1
playwright==1.48.0
beautifulsoup4==4.12.3
PyMuPDF==1.24.11
chromadb==0.5.15
tenacity==9.0.0
python-dotenv==1.0.1
